refactor: 분리된 useMainSection 훅 합치기
